Update checkers, links, photo load, DB defaults

+++ src/DocsMan.App/Interactors/UserExec.cs
@@ -205,6 +205,9 @@ namespace DocsMan.App.Interactors
 		{
 			try
 			{
+				if (roleId == 1 || (roleId == 2 && userId == 1))
+					return new("Запрещено удалять эту роль", "Forbidden delete this role");
+				
 				await _userRoles.DeleteBindAsync(
 					new()
 					{
